# repo-updater: Improve performance when calculating notClonedCount