--- a/cmd/gitserver/server/server.go
+++ b/cmd/gitserver/server/server.go
@@ -232,6 +232,7 @@ func (s *Server) Handler() http.Handler {
 	mux.HandleFunc("/repo-update", s.handleRepoUpdate)
 	mux.HandleFunc("/getGitolitePhabricatorMetadata", s.handleGetGitolitePhabricatorMetadata)
 	mux.HandleFunc("/create-commit-from-patch", s.handleCreateCommitFromPatch)
+	mux.HandleFunc("/cloned-count", s.handleClonedCount)
 	mux.HandleFunc("/ping", func(w http.ResponseWriter, _ *http.Request) {
 		w.WriteHeader(http.StatusOK)
 	})
@@ -1272,6 +1273,25 @@ func (s *Server) ensureRevision(ctx context.Context, repo api.RepoName, url, rev
 	// Revision not found, update before returning.
 	s.doRepoUpdate(ctx, repo, url)
 	return true
+}
+
+func (s *Server) handleClonedCount(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	var resp protocol.ClonedCountResponse
+
+	err := s.walkCloned(ctx, func(path string) error {
+		resp.Count++
+		return nil
+	})
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	if err := json.NewEncoder(w).Encode(resp); err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
 }
 
 // quickRevParseHead best-effort mimics the execution of `git rev-parse HEAD`, but doesn't exec a child process.