--- a/cmd/repo-updater/repos/store.go
+++ b/cmd/repo-updater/repos/store.go
@@ -27,7 +27,7 @@ type Store interface {
 	ListRepos(context.Context, StoreListReposArgs) ([]*Repo, error)
 	UpsertRepos(ctx context.Context, repos ...*Repo) error
 
-	ListAllRepoNames(context.Context) ([]api.RepoName, error)
+	CountRepos(context.Context) (int, error)
 }
 
 // StoreListReposArgs is a query arguments type used by
@@ -369,37 +369,28 @@ func listReposQuery(args StoreListReposArgs) paginatedQuery {
 	}
 }
 
-// ListAllRepoNames lists the names of all stored repos
-func (s DBStore) ListAllRepoNames(ctx context.Context) (names []api.RepoName, _ error) {
-	return names, s.paginate(ctx, 0, 0, listAllRepoNamesQuery,
-		func(sc scanner) (last, count int64, err error) {
-			var (
-				id   int64
-				name api.RepoName
-			)
-			if err = sc.Scan(&id, &name); err != nil {
-				return 0, 0, err
-			}
-			names = append(names, name)
-			return id, 1, nil
-		},
-	)
+// CountRepos lists the names of all stored repos
+func (s DBStore) CountRepos(ctx context.Context) (count int, err error) {
+	rows, err := s.db.QueryContext(ctx, countReposQuery)
+	if err != nil {
+		return count, errors.Wrap(err, "CountRepos")
+	}
+
+	defer closeErr(rows, &err)
+
+	rows.Next()
+
+	err = rows.Scan(&count)
+	return count, err
 }
 
-const listAllRepoNamesQueryFmtstr = `
--- source: cmd/repo-updater/repos/store.go:DBStore.ListAllRepoNames
+const countReposQuery = `
+-- source: cmd/repo-updater/repos/store.go:DBStore.CountRepos
 SELECT
-  id,
-  name
+  COUNT(*)
 FROM repo
-WHERE id > %s
-AND deleted_at IS NULL
-ORDER BY id ASC LIMIT %s
+WHERE deleted_at IS NULL
 `
-
-func listAllRepoNamesQuery(cursor, limit int64) *sqlf.Query {
-	return sqlf.Sprintf(listAllRepoNamesQueryFmtstr, cursor, limit)
-}
 
 // a paginatedQuery returns a query with the given pagination
 // parameters