--- a/cmd/repo-updater/repos/testing.go
+++ b/cmd/repo-updater/repos/testing.go
@@ -56,7 +56,7 @@ type FakeStore struct {
 	GetRepoByNameError          error // error to be returned in GetRepoByName
 	ListReposError              error // error to be returned in ListRepos
 	UpsertReposError            error // error to be returned in UpsertRepos
-	ListAllRepoNamesError       error // error to be returned in ListAllRepoNames
+	CountReposError             error // error to be returned in CountRepos
 
 	svcIDSeq  int64
 	repoIDSeq uint32
@@ -88,7 +88,7 @@ func (s *FakeStore) Transact(ctx context.Context) (TxStore, error) {
 		GetRepoByNameError:          s.GetRepoByNameError,
 		ListReposError:              s.ListReposError,
 		UpsertReposError:            s.UpsertReposError,
-		ListAllRepoNamesError:       s.ListAllRepoNamesError,
+		CountReposError:             s.CountReposError,
 
 		svcIDSeq:  s.svcIDSeq,
 		svcByID:   svcByID,
@@ -251,18 +251,13 @@ func (s FakeStore) ListRepos(ctx context.Context, args StoreListReposArgs) ([]*R
 	return repos, nil
 }
 
-// ListAllRepoNames lists names of all repos in the store
-func (s FakeStore) ListAllRepoNames(ctx context.Context) ([]api.RepoName, error) {
-	if s.ListAllRepoNamesError != nil {
-		return nil, s.ListAllRepoNamesError
+// CountRepos lists names of all repos in the store
+func (s FakeStore) CountRepos(ctx context.Context) (int, error) {
+	if s.CountReposError != nil {
+		return 0, s.CountReposError
 	}
 
-	names := make([]api.RepoName, 0, len(s.repoByID))
-	for _, r := range s.repoByID {
-		names = append(names, api.RepoName(r.Name))
-	}
-
-	return names, nil
+	return len(s.repoByID), nil
 }
 
 func evalOr(bs ...bool) bool {