--- a/cmd/repo-updater/repoupdater/server_test.go
+++ b/cmd/repo-updater/repoupdater/server_test.go
@@ -23,7 +23,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/github"
 	"github.com/sourcegraph/sourcegraph/pkg/extsvc/gitlab"
 	"github.com/sourcegraph/sourcegraph/pkg/gitserver"
-	gitprotocol "github.com/sourcegraph/sourcegraph/pkg/gitserver/protocol"
 	"github.com/sourcegraph/sourcegraph/pkg/jsonc"
 	"github.com/sourcegraph/sourcegraph/pkg/repoupdater"
 	"github.com/sourcegraph/sourcegraph/pkg/repoupdater/protocol"
@@ -546,24 +545,44 @@ func TestServer_RepoExternalServices(t *testing.T) {
 func TestServer_StatusMessages(t *testing.T) {
 	testCases := []struct {
 		name              string
-		gitserverResponse *gitprotocol.AreReposClonedResponse
+		stored            repos.Repos
+		gitserverResponse []string
 		res               *protocol.StatusMessagesResponse
 		err               string
 	}{
 		{
-			name: "nothing cloning",
-			gitserverResponse: &gitprotocol.AreReposClonedResponse{
-				Results: map[api.RepoName]bool{"foobar": true},
-			},
+			name:              "all cloned",
+			stored:            []*repos.Repo{{Name: "foobar"}},
+			gitserverResponse: []string{"foobar"},
 			res: &protocol.StatusMessagesResponse{
 				Messages: []protocol.StatusMessage{},
 			},
 		},
 		{
-			name: "repositories cloning",
-			gitserverResponse: &gitprotocol.AreReposClonedResponse{
-				Results: map[api.RepoName]bool{"foobar": false, "barfoo": false, "barbaz": false},
+			name:              "more cloned than in DB",
+			stored:            []*repos.Repo{{Name: "foobar"}},
+			gitserverResponse: []string{"foobar", "barfoo"},
+			res: &protocol.StatusMessagesResponse{
+				Messages: []protocol.StatusMessage{},
 			},
+		},
+		{
+			name:              "few not cloned",
+			stored:            []*repos.Repo{{Name: "foobar"}, {Name: "barfoo"}, {Name: "barbaz"}},
+			gitserverResponse: []string{"foobar"},
+			res: &protocol.StatusMessagesResponse{
+				Messages: []protocol.StatusMessage{
+					{
+						Type:    protocol.CloningStatusMessage,
+						Message: "2 repositories enqueued for cloning...",
+					},
+				},
+			},
+		},
+		{
+			name:              "nothing cloned",
+			stored:            []*repos.Repo{{Name: "foobar"}, {Name: "barfoo"}, {Name: "barbaz"}},
+			gitserverResponse: []string{},
 			res: &protocol.StatusMessagesResponse{
 				Messages: []protocol.StatusMessage{
 					{
@@ -580,8 +599,14 @@ func TestServer_StatusMessages(t *testing.T) {
 		ctx := context.Background()
 
 		t.Run(tc.name, func(t *testing.T) {
-			gitserverClient := &fakeGitserverClient{areReposClonedResponse: tc.gitserverResponse}
+			gitserverClient := &fakeGitserverClient{listClonedResponse: tc.gitserverResponse}
+
 			store := new(repos.FakeStore)
+			err := store.UpsertRepos(ctx, tc.stored.Clone()...)
+			if err != nil {
+				t.Fatal(err)
+			}
+
 			s := &Server{Store: store, GitserverClient: gitserverClient}
 
 			srv := httptest.NewServer(s.Handler())
@@ -936,11 +961,11 @@ func (s *fakeScheduler) ScheduleInfo(id uint32) *protocol.RepoUpdateSchedulerInf
 }
 
 type fakeGitserverClient struct {
-	areReposClonedResponse *gitprotocol.AreReposClonedResponse
+	listClonedResponse []string
 }
 
-func (g *fakeGitserverClient) AreReposCloned(ctx context.Context, names ...api.RepoName) (*gitprotocol.AreReposClonedResponse, error) {
-	return g.areReposClonedResponse, nil
+func (g *fakeGitserverClient) ListCloned(ctx context.Context) ([]string, error) {
+	return g.listClonedResponse, nil
 }
 
 func formatJSON(s string) string {