@mrnugget 2019-07-05
I realized yesterday that this PR introduced a possible bug. I'm quoting myself on Zulip here:

> if the list of cloned repos that gitserver reports contains repos that have already been deleted from repo-updater's repo table, then we probably need to revert my PR from yesterday :frown: The possible bug is this: you remove an external with say 500 repos and then add a new one with 500 _other_ repos immediately after. If gitserver then still reports "500 cloned repos" repo-updater will say "I have 500 repos, so the diff is 0, which means 0 repos need to be cloned" Of course, it'll be eventually consistent, but still...

@keegancsmith confirmed this:

> yeah we leave the clones in place, so you are right we can't rely on just the count. We need to do some diffing :/

But I think there is a possibility to speed up the original version, that includes diffing, by not sending the complete list of repos over to gitserver, but instead doing the diffing locally.

What I'll do is

1.) Revert this PR
2.) Investigate whether a combination of this new idea and the `godirwalk` optimization in https://github.com/sourcegraph/sourcegraph/pull/4802 gives us better performance than the original version